Migrate utility functions to util
